# pkg/storage/metastore: Optimize SQL query building of locations by ids