--- a/pkg/storage/metastore/sql.go
+++ b/pkg/storage/metastore/sql.go
@@ -194,6 +194,7 @@ func (s *sqlMetaStore) GetLocationsByIDs(ctx context.Context, ids ...uint64) (
 	mappingIDsSeen := map[uint64]struct{}{}
 
 	remainingIds := []uint64{}
+	maxRemainingID := uint64(0)
 	for _, id := range ids {
 		l, found, err := s.cache.getLocationByID(ctx, id)
 		if err != nil {
@@ -209,12 +210,15 @@ func (s *sqlMetaStore) GetLocationsByIDs(ctx context.Context, ids ...uint64) (
 			}
 			continue
 		}
+		if maxRemainingID < id {
+			maxRemainingID = id
+		}
 		remainingIds = append(remainingIds, id)
 	}
 
 	if len(remainingIds) > 0 {
 		dbctx, dbspan := s.tracer.Start(ctx, "GetLocationsByIDs-SQL-query")
-		rows, err := s.db.QueryContext(dbctx, buildLocationsByIDsQuery(remainingIds))
+		rows, err := s.db.QueryContext(dbctx, buildLocationsByIDsQuery(maxRemainingID, remainingIds))
 		dbspan.End()
 		if err != nil {
 			return nil, fmt.Errorf("execute SQL query: %w", err)
@@ -302,18 +306,34 @@ func (s *sqlMetaStore) GetLocationsByIDs(ctx context.Context, ids ...uint64) (
 	return res, nil
 }
 
-func buildLocationsByIDsQuery(ids []uint64) string {
-	sIds := ""
-	for i, id := range ids {
+const (
+	locsByIDsQueryStart = `SELECT "id", "mapping_id", "address", "is_folded", "normalized_address", "lines"
+				FROM "locations"
+				WHERE id IN (`
+)
+
+func buildLocationsByIDsQuery(max uint64, ids []uint64) string {
+	maxLen := len(strconv.FormatUint(max, 10))
+
+	query := make([]byte, 0,
+		// The max value is known, and invididual string can be larger than it.
+		len(ids)*maxLen+
+			// Add the start of the query.
+			len(locsByIDsQueryStart)+
+			// len(ids)-1 commas, and a closing bracket is len(ids).
+			len(ids),
+	)
+	query = append(query, locsByIDsQueryStart...)
+
+	for i := range ids {
 		if i > 0 {
-			sIds += ","
+			query = append(query, comma)
 		}
-		sIds += strconv.FormatInt(int64(id), 10)
+		query = strconv.AppendUint(query, ids[i], 10)
 	}
 
-	return fmt.Sprintf(`SELECT "id", "mapping_id", "address", "is_folded", "normalized_address", "lines"
-				FROM "locations"
-				WHERE id IN (%s)`, sIds)
+	query = append(query, closingBracket)
+	return unsafeString(query)
 }
 
 func (s *sqlMetaStore) GetMappingsByIDs(ctx context.Context, ids ...uint64) (map[uint64]*profile.Mapping, error) {
@@ -466,7 +486,7 @@ func (s *sqlMetaStore) getLinesByLocationIDs(ctx context.Context, ids ...uint64)
 }
 
 const (
-	queryStart = `SELECT "location_id", "line", "function_id"
+	linesByLocationsIDsQueryStart = `SELECT "location_id", "line", "function_id"
 			FROM "lines" WHERE location_id IN (`
 	comma          = ','
 	closingBracket = ')'
@@ -479,11 +499,11 @@ func buildLinesByLocationIDsQuery(max uint64, ids []uint64) string {
 		// The max value is known, and invididual string can be larger than it.
 		len(ids)*maxLen+
 			// Add the start of the query.
-			len(queryStart)+
+			len(linesByLocationsIDsQueryStart)+
 			// len(ids)-1 commas, and a closing bracket is len(ids).
 			len(ids),
 	)
-	query = append(query, queryStart...)
+	query = append(query, linesByLocationsIDsQueryStart...)
 
 	for i := range ids {
 		if i > 0 {