--- a/pkg/storage/metastore/sql_bench_test.go
+++ b/pkg/storage/metastore/sql_bench_test.go
@@ -47,7 +47,7 @@ func BenchmarkBuildLocationsByIDsQuery(b *testing.B) {
 			}
 
 			for i := 0; i < b.N; i++ {
-				result = buildLocationsByIDsQuery(input)
+				result = buildLocationsByIDsQuery(input[len(input)-1], input)
 			}
 		})
 	}