This often shows up in large queries of previously unqueried data.

```
name                               old time/op    new time/op    delta
BuildLocationsByIDsQuery/1-8          378ns ± 0%      89ns ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/10-8        1.42µs ± 0%    0.17µs ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/100-8       20.8µs ± 0%     0.9µs ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/1000-8      1.46ms ± 0%    0.02ms ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/10000-8      129ms ± 0%       0ms ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/100000-8     12.5s ± 0%      0.0s ± 0%   ~     (p=1.000 n=1+1)

name                               old alloc/op   new alloc/op   delta
BuildLocationsByIDsQuery/1-8           144B ± 0%      128B ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/10-8          392B ± 0%      144B ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/100-8       30.0kB ± 0%     0.4kB ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/1000-8      4.07MB ± 0%    0.00MB ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/10000-8      517MB ± 0%       0MB ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/100000-8    58.7GB ± 0%     0.0GB ± 0%   ~     (p=1.000 n=1+1)

name                               old allocs/op  new allocs/op  delta
BuildLocationsByIDsQuery/1-8           2.00 ± 0%      1.00 ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/10-8          20.0 ± 0%       1.0 ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/100-8          200 ± 0%         1 ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/1000-8       2.90k ± 0%     0.00k ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/10000-8      29.9k ± 0%      0.0k ± 0%   ~     (p=1.000 n=1+1)
BuildLocationsByIDsQuery/100000-8      301k ± 0%        0k ± 0%   ~     (p=1.000 n=1+1)
```